requirements.txt
langchain
langchain-community
langchain-openai
pgvector
python-dotenv
psycopg2-binary
pypdf
weasyprint
Markdown
pandas
azure-storage-blob
langchain-ai21
unstructured[all-docs]
streamlit
llama-index
llama-parse
pinecone-client
pinecone-text
bs4

load-dotenv
langchain-huggingface